ADCS/RW:  Moved v0.7 PID and other assorted code into new rw.c/.h library, in a new dsbase/actuators folder.  Includes a number of fixes to get us to around +/-5rpm with very little overshoot or oscillation, though more work is needed, both in terms of tuning and mitigating other issues (like integrative term windup, which can still be bad, especially if the motor stalls at the beginning for whatever reason and the integral term gets massive).  NO support for crossing zero into "negative RPM" yet, need to talk to RAIN lab guys about most appropriate way to do that.

+++ src/dsbase/core/timers.h
@@ -12,6 +12,6 @@
 
 #define SEC   8000000   // Ideal number of cycles/second
 
-uint16_t timerCycleDiff16(uint16_t earlierTime, uint16_t laterTime);
+uint32_t timerCycleDiff16(uint16_t earlierTime, uint16_t laterTime);
 
 #endif /* DSBASE_CORE_TIMERS_H_ */
